refactor: 비슷한 파일 racingGame 폴더로 분리, 자주쓰는 함수 utils 파일에 분리
